Share the Gemini edit/write_file diff rendering
Both observations rendered a created or changed file the same way: the created line, the diff via visualize_diff, and the same _diff_cache private attr. The only real difference was the wording of the line for an existing file.

Move the rendering to a FileChangeObservation base and let each tool supply that one line through change_summary(). The base extends ABC so it stays out of the discriminated union and out of the OpenAPI schema, and both tools keep declaring their own fields so their descriptions and field order are untouched.

Part of #4224
