+++ openhands-tools/openhands/tools/gemini/edit/definition.py
@@ -4,17 +4,16 @@ from collections.abc import Sequence
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from pydantic import Field, PrivateAttr
-from rich.text import Text
+from pydantic import Field
 
 from openhands.sdk.tool import (
     Action,
     DeclaredResources,
-    Observation,
     ToolAnnotations,
     ToolDefinition,
     register_tool,
 )
+from openhands.tools.gemini.file_change import FileChangeObservation
 
 
 if TYPE_CHECKING:
@@ -43,7 +42,7 @@ class EditAction(Action):
     )
 
 
-class EditObservation(Observation):
+class EditObservation(FileChangeObservation):
     """Observation from editing a file."""
 
     file_path: str | None = Field(
@@ -62,45 +61,8 @@ class EditObservation(Observation):
         default=None, description="The content after the edit."
     )
 
-    _diff_cache: Text | None = PrivateAttr(default=None)
-
-    @property
-    def visualize(self) -> Text:
-        """Return Rich Text representation of this observation."""
-        text = Text()
-
-        if self.is_error:
-            text.append("❌ ", style="red bold")
-            text.append(self.ERROR_MESSAGE_HEADER, style="bold red")
-            return super().visualize
-
-        if self.file_path:
-            if self.is_new_file:
-                text.append("✨ ", style="green bold")
-                text.append(f"Created: {self.file_path}\n", style="green")
-            else:
-                text.append("✏️  ", style="yellow bold")
-                text.append(
-                    (
-                        f"Edited: {self.file_path} "
-                        f"({self.replacements_made} replacement(s))\n"
-                    ),
-                    style="yellow",
-                )
-
-            if self.old_content is not None and self.new_content is not None:
-                from openhands.tools.file_editor.utils.diff import visualize_diff
-
-                if not self._diff_cache:
-                    self._diff_cache = visualize_diff(
-                        self.file_path,
-                        self.old_content,
-                        self.new_content,
-                        n_context_lines=2,
-                        change_applied=True,
-                    )
-                text.append(self._diff_cache)
-        return text
+    def change_summary(self) -> str:
+        return f"Edited: {self.file_path} ({self.replacements_made} replacement(s))\n"
 
 
 TOOL_DESCRIPTION = """Replaces text within a file.
