+++ openhands-tools/openhands/tools/gemini/write_file/definition.py
@@ -4,17 +4,16 @@ from collections.abc import Sequence
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from pydantic import Field, PrivateAttr
-from rich.text import Text
+from pydantic import Field
 
 from openhands.sdk.tool import (
     Action,
     DeclaredResources,
-    Observation,
     ToolAnnotations,
     ToolDefinition,
     register_tool,
 )
+from openhands.tools.gemini.file_change import FileChangeObservation
 
 
 if TYPE_CHECKING:
@@ -28,7 +27,7 @@ class WriteFileAction(Action):
     content: str = Field(description="The content to write to the file.")
 
 
-class WriteFileObservation(Observation):
+class WriteFileObservation(FileChangeObservation):
     """Observation from writing a file."""
 
     file_path: str | None = Field(
@@ -44,39 +43,8 @@ class WriteFileObservation(Observation):
         default=None, description="The new content written to the file."
     )
 
-    _diff_cache: Text | None = PrivateAttr(default=None)
-
-    @property
-    def visualize(self) -> Text:
-        """Return Rich Text representation of this observation."""
-        text = Text()
-
-        if self.is_error:
-            text.append("❌ ", style="red bold")
-            text.append(self.ERROR_MESSAGE_HEADER, style="bold red")
-            return super().visualize
-
-        if self.file_path:
-            if self.is_new_file:
-                text.append("✨ ", style="green bold")
-                text.append(f"Created: {self.file_path}\n", style="green")
-            else:
-                text.append("✏️  ", style="yellow bold")
-                text.append(f"Updated: {self.file_path}\n", style="yellow")
-
-            if self.old_content is not None and self.new_content is not None:
-                from openhands.tools.file_editor.utils.diff import visualize_diff
-
-                if not self._diff_cache:
-                    self._diff_cache = visualize_diff(
-                        self.file_path,
-                        self.old_content,
-                        self.new_content,
-                        n_context_lines=2,
-                        change_applied=True,
-                    )
-                text.append(self._diff_cache)
-        return text
+    def change_summary(self) -> str:
+        return f"Updated: {self.file_path}\n"
 
 
 TOOL_DESCRIPTION = """Writes content to a specified file in the local filesystem.
